Med: stonith-helper: improve to be able to dynamically generate a waiting time.
It is possible to specifies the external script.
This feature use the return value of the script as a wait time.

User can specify a any acript.
However, the return code must be 0 at the normal time.
And the standard output must be a waiting time(seconds).

I added two parameters below.

[run_variable_wait]
If it is not "no", run the variable_wait_script and wait for that return.
If it is "no" or not set, Don't wait.

[variable_wait_script]
Path of the external script.
If return code of command is "0" and return value greater than 0, this node wait for the return value[seconds].